Initial commit: screen and widget modules
