+++ src/app/widget_modules/monitor/monitor.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { SharedModule } from '../../app/shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [],
@@ -7,4 +7,4 @@ import { SharedModule } from '../../app/shared/shared.module';
     SharedModule
   ]
 })
-export class PerformanceTestsModule { }
+export class MonitorModule { }
